Move initialize function for Gravity gradient
