+++ src/disturbances/gravity_gradient.cpp
@@ -7,6 +7,7 @@
 
 #include <cmath>
 #include <environment/global/physical_constants.hpp>
+#include <library/initialize/initialize_file_access.hpp>
 
 #include "../library/logger/log_utility.hpp"
 
@@ -47,4 +48,26 @@ std::string GravityGradient::GetLogValue() const {
   str_tmp += WriteVector(torque_b_Nm_);
 
   return str_tmp;
+}
+
+GravityGradient InitGravityGradient(const std::string initialize_file_path) {
+  auto conf = IniAccess(initialize_file_path);
+  const char* section = "GRAVITY_GRADIENT";
+
+  const bool is_calc_enable = conf.ReadEnable(section, INI_CALC_LABEL);
+  GravityGradient gg_disturbance(is_calc_enable);
+  gg_disturbance.is_log_enabled_ = conf.ReadEnable(section, INI_LOG_LABEL);
+
+  return gg_disturbance;
+}
+
+GravityGradient InitGravityGradient(const std::string initialize_file_path, const double gravity_constant_m3_s2) {
+  auto conf = IniAccess(initialize_file_path);
+  const char* section = "GRAVITY_GRADIENT";
+
+  const bool is_calc_enable = conf.ReadEnable(section, INI_CALC_LABEL);
+  GravityGradient gg_disturbance(gravity_constant_m3_s2, is_calc_enable);
+  gg_disturbance.is_log_enabled_ = conf.ReadEnable(section, INI_LOG_LABEL);
+
+  return gg_disturbance;
 }
