+++ src/disturbances/initialize_disturbances.hpp
@@ -15,21 +15,6 @@
 #include <disturbances/third_body_gravity.hpp>
 
 /**
- * @fn InitGravityGradient
- * @brief Initialize GravityGradient class with earth gravitational constant
- * @param [in] initialize_file_path: Initialize file path
- */
-GravityGradient InitGravityGradient(const std::string initialize_file_path);
-
-/**
- * @fn InitGravityGradient
- * @brief Initialize GravityGradient class with earth gravitational constant
- * @param [in] initialize_file_path: Initialize file path
- * @param [in] gravity_constant_m3_s2: Gravitational constant [m3/s2]
- */
-GravityGradient InitGravityGradient(const std::string initialize_file_path, const double gravity_constant_m3_s2);
-
-/**
  * @fn InitMagneticDisturbance
  * @brief Initialize MagneticDisturbance class with earth gravitational constant
  * @param [in] initialize_file_path: Initialize file path
